+++ src/commands/api/getAzureContainerAppsApiProvider.ts
@@ -4,6 +4,7 @@
 *--------------------------------------------------------------------------------------------*/
 
 import { createApiProvider, type apiUtils } from "@microsoft/vscode-azext-utils";
+import { deployImageApi } from "./deployImageApi";
 import { deployWorkspaceProjectApi } from "./deployWorkspaceProjectApi";
 import type * as api from "./vscode-azurecontainerapps.api";
 
@@ -12,6 +13,7 @@ export function getAzureContainerAppsApiProvider(): apiUtils.AzureExtensionApiPr
         // Todo: Change this to 0.0.3 later.  0.0.3 is backwards compatible anyway so this change should be fine either way.
         // For some reason it's causing a block on Function side, so just keep it at 0.0.1 until we figure out why
         apiVersion: '0.0.1',
-        deployWorkspaceProject: deployWorkspaceProjectApi
+        deployImage: deployImageApi,
+        deployWorkspaceProject: deployWorkspaceProjectApi,
     }]);
 }